Update 'libretro_core_options.h' to v1.3
